refactorizado de la estructura de ficheros, removed dupes
