--- server/src/WeatherService.proto
@@ -1,14 +1,0 @@
-syntax = "proto3";
-
-package weather;
-
-service WeatherService {
-    rpc GetWeather(GetWeatherRequest) returns (Weather);
-}
-message GetWeatherRequest {
-    string city = 1;
-}
-message Weather {
-    string city = 1;
-    string weather = 2;
-}
