+++ src/common/WeatherService.proto
