+++ src/server/static/client.js
