--- weatherservice/src/dirname.js
@@ -1,4 +1,0 @@
-import { fileURLToPath } from 'url';
-import { dirname } from 'path';
-
-export const __dirname = dirname(fileURLToPath(import.meta.url));
